move CLI definition to FSORepairTool
